+++ src/definitions/behaviors/form.js
@@ -946,7 +946,7 @@ $.fn.form.settings = {
   namespace         : 'form',
 
   debug             : false,
-  verbose           : true,
+  verbose           : false,
   performance       : true,
 
   fields            : false,